Fix fast refresh compatibility for document templates
