+++ frontend/src/employee-frontend/components/document-templates/template-editor/TemplateModal.tsx
@@ -5,34 +5,11 @@
 import React, { useMemo } from 'react'
 
 import DateRange from 'lib-common/date-range'
-import {
-  boolean,
-  openEndedLocalDateRange,
-  string
-} from 'lib-common/form/fields'
-import {
-  array,
-  object,
-  oneOf,
-  required,
-  transformed,
-  validated,
-  value
-} from 'lib-common/form/form'
+import { openEndedLocalDateRange } from 'lib-common/form/fields'
 import { useForm, useFormFields } from 'lib-common/form/hooks'
-import { ValidationError, ValidationSuccess } from 'lib-common/form/types'
-import { nonBlank } from 'lib-common/form/validators'
-import type {
-  DocumentTemplateBasicsRequest,
-  DocumentType,
-  ExportedDocumentTemplate
-} from 'lib-common/generated/api-types/document'
+import type { ExportedDocumentTemplate } from 'lib-common/generated/api-types/document'
 import { documentTypes } from 'lib-common/generated/api-types/document'
-import type { PlacementType } from 'lib-common/generated/api-types/placement'
-import type {
-  DocumentTemplateId,
-  UiLanguage
-} from 'lib-common/generated/api-types/shared'
+import type { DocumentTemplateId } from 'lib-common/generated/api-types/shared'
 import { uiLanguages } from 'lib-common/generated/api-types/shared'
 import type { JsonOf } from 'lib-common/json'
 import { useMutationResult } from 'lib-common/query'
@@ -51,100 +28,12 @@ import {
 } from 'lib-customizations/employee'
 
 import { useTranslation } from '../../../state/i18n'
+import { documentTemplateForm } from '../forms'
 import {
   createDocumentTemplateMutation,
   duplicateDocumentTemplateMutation,
   importDocumentTemplateMutation
 } from '../queries'
-
-export const documentTemplateForm = transformed(
-  object({
-    name: validated(string(), nonBlank),
-    type: required(oneOf<DocumentType>()),
-    placementTypes: validated(array(value<PlacementType>()), (arr) =>
-      arr.length === 0 ? 'required' : undefined
-    ),
-    language: required(oneOf<UiLanguage>()),
-    confidential: boolean(),
-    confidentialityDurationYears: required(value<string>()),
-    confidentialityBasis: required(value<string>()),
-    legalBasis: string(),
-    validity: required(openEndedLocalDateRange()),
-    processDefinitionNumber: required(value<string>()),
-    archiveDurationMonths: required(value<string>()),
-    archiveExternally: boolean()
-  }),
-  (value) => {
-    const archived = value.processDefinitionNumber.trim().length > 0
-    if (archived) {
-      const archiveDurationMonths = parseInt(value.archiveDurationMonths)
-      if (isNaN(archiveDurationMonths) || archiveDurationMonths < 1) {
-        return ValidationError.field('archiveDurationMonths', 'integerFormat')
-      }
-    }
-
-    if (value.archiveExternally) {
-      if (value.processDefinitionNumber.trim().length === 0) {
-        return ValidationError.field('processDefinitionNumber', 'required')
-      }
-
-      if (value.archiveDurationMonths.trim().length === 0) {
-        return ValidationError.field('archiveDurationMonths', 'required')
-      }
-
-      const archiveDurationMonths = parseInt(value.archiveDurationMonths)
-      if (isNaN(archiveDurationMonths) || archiveDurationMonths < 1) {
-        return ValidationError.field('archiveDurationMonths', 'integerFormat')
-      }
-    }
-
-    const confidential = value.confidential
-    if (confidential) {
-      const confidentialityDurationYears = parseInt(
-        value.confidentialityDurationYears
-      )
-      if (
-        isNaN(confidentialityDurationYears) ||
-        confidentialityDurationYears < 1
-      ) {
-        return ValidationError.field(
-          'confidentialityDurationYears',
-          'integerFormat'
-        )
-      }
-      if (value.confidentialityBasis.trim().length === 0) {
-        return ValidationError.field('confidentialityBasis', 'required')
-      }
-    }
-
-    const output: DocumentTemplateBasicsRequest = {
-      ...value,
-
-      confidentiality: confidential
-        ? {
-            durationYears: parseInt(value.confidentialityDurationYears),
-            basis: value.confidentialityBasis.trim()
-          }
-        : null,
-      ...(value.archiveExternally
-        ? {
-            templateType: 'ARCHIVED_EXTERNALLY',
-            processDefinitionNumber: value.processDefinitionNumber.trim(),
-            archiveDurationMonths: parseInt(value.archiveDurationMonths)
-          }
-        : {
-            templateType: 'REGULAR',
-            processDefinitionNumber: archived
-              ? value.processDefinitionNumber.trim()
-              : null,
-            archiveDurationMonths: archived
-              ? parseInt(value.archiveDurationMonths)
-              : null
-          })
-    }
-    return ValidationSuccess.of(output)
-  }
-)
 
 export type TemplateModalMode =
   | { type: 'new' }
