+++ frontend/src/employee-frontend/components/document-templates/template-editor/TemplateSectionView.tsx
@@ -20,7 +20,7 @@ import colors from 'lib-customizations/common'
 import { faArrowDown, faArrowUp, faPen, faPlus, faTrash } from 'lib-icons'
 
 import { useTranslation } from '../../../state/i18n'
-import { type templateSectionForm } from '../templates'
+import type { templateSectionForm } from '../forms'
 
 import TemplateQuestionModal from './TemplateQuestionModal'
 import TemplateQuestionView from './TemplateQuestionView'
